+++ src/commonMain/kotlin/dev/inmo/saucenaoapi/models/LimitsState.kt
@@ -8,4 +8,6 @@ data class LimitsState(
     val maxLongQuota: Int,
     val knownShortQuota: Int,
     val knownLongQuota: Int
-)
+) : Comparable<LimitsState> {
+    override fun compareTo(other: LimitsState): Int = knownShortQuota.compareTo(other.knownShortQuota)
+}